updat orm

+++ src/main/java/com/codegym/model/SongForm.java
@@ -3,7 +3,7 @@ package com.codegym.model;
 import org.springframework.web.multipart.MultipartFile;
 
 public class SongForm {
-    private int id;
+    private Long id;
 
     private String name;
 
@@ -16,7 +16,7 @@ public class SongForm {
     public SongForm() {
     }
 
-    public SongForm(int id, String name, String singer, String type, MultipartFile songfile) {
+    public SongForm(Long id, String name, String singer, String type, MultipartFile songfile) {
         this.id = id;
         this.name = name;
         this.singer = singer;
@@ -24,11 +24,11 @@ public class SongForm {
         this.songfile = songfile;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
